deploy/cli/pyproject.toml
[project]
name = "skill-cli"
version = "1.0.0"
description = "红龙获客系统技能包管理工具链"
requires-python = ">=3.8"
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml"]

[project.scripts]
skill-cli = "skill_cli:main"
